requirements.txt
certifi==2019.3.9
chardet==3.0.4
config==0.4.0
confusable-homoglyphs==3.2.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-heroku==0.3.1
django-model-utils==3.1.2
django-ratings==0.3.7
django-registration==2.4.1
django-star-ratings==0.7.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-core==3.1.0
swapper==1.1.0
urllib3==1.24.1
userprofile==0.2.5
whitenoise==4.0
